In this commit, at least the reason for which after 16 seconds (i.e. STA start to scan other channels), the NIC of STA seems to be down is idenfied:
In ns-3.22, SetChannelNumner is a memember method of WifiPhy class but is
actually implemented in its subclass YansWifiPhy class. The implementation
of this method considers different cases (modelized by an enum number in WifiPhy
class): IDLE, TX, RX, etc.

However, in ns-3.27, SetChannelNumber is directly but significantly different
than that in YansWifiPhy class of ns-3.22.

Two solutions are thus available:
1) Copy the old implementation of YansWifiPhy ns-3.22 to the same class of
ns-3.27.
2) Adaptation after understanding the implementation of ns-3.27