To implement TCP/UDP throughput estimation, we have to use ns3 flow monitor module, which requires some modification jobs to this project. During this work, some new simulation scripts and bugs are found.
In this commit, we add TCP bulk send application in scripts such as lisp_mobility_between_subnet and lisp_mobility_between_subnet_single_encap.
A new script called lisp_reference is added which does not consider LISP mobility, does not use Wifi neither DHCP. It only considers the basic
LISP and is used to help understand and master flow monitor in the context of LISP.

Some bugs are found and fixed in this commit:
1) Outer IP header and UDP header length field calculation in LispOutput method
	In ns-3, the m_payload attribute in UDP/IP header refers to the payload size as what its name indicates. Through the generated PCAP files,
we found that outer IP header and UDP header of encapsulated LISP packets have wrong length. This bug is fixed.
2) Incorrect UDP length after LISP decapsulation
	Still by PCAP files, we find the UDP content is always 8 bytes less than what it shoule be. Surprisingly, before removing the UDP header
from the packet, UDP has the correct length field. Even in the peeled UDP header, m_payload is still correct. The problems comes from the operation
of adding the removed UDP header to the packet. I doubt this bug comes from packet addHeader or serialized method, but still find the exact cause.
As a workaround, during LISPInput method, insteading of removing UDP header, I copy the packet and remove the UDP header and do the check jobs on the
copied packet. This solution works well.
3) No check for if the smart pointer to EID is null in DHCP client
	When working on the lisp-reference script, I was stuck by a segmentation fault. After investigation, I found the bug: for a device without TUN
device. The smart pointer to EID is surely 0. However, I have no check for this before.

Still a few of problems to be solved:
1) Investigate why with TUN device, each packet transmitted by LISP-MN is duplicated. For UDP case, no big deal. For TCP case, TCP-out-of-order and
TCP retransmission are observed, which maybe has impact on throughput estimation.
2) DHCP client always sent three DHCP Discovery and DHCP Request three times. I think the two supplementary messages are due to wifi channel and TUN
device. This remains as a problem for a long time.